feat: create model-entitys

+++ src/main/java/br/com/meli/desafio_final/model/entity/Adsense.java
@@ -4,5 +4,5 @@ public class Adsense {
     long id;
     Seller seller;
     Product product;
-    double price;
+    float price;
 }
